frameserver/requirements.txt
flask
google-cloud-storage
gunicorn
